+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Add a DMA
 
 ### Fixed
+- Measure the `wfi` stalls and stalls caused by `opc` properly
 
 ### Changed
 - Increase the default AXI width